+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## NEXT VERSION
 
+- fix: header row's height doesn't update on `headerHeight` change
+
 ## v1.0.0 (2019-03-26)
 
 Initial public release